debugging

+++ app/controllers/posts_controller.rb
@@ -1,23 +1,35 @@
 class PostsController < ApplicationController
-  skip_before_action :authorize, only: [:create]
+  skip_before_action :authorize, only: [:index, :show]
 
   def index
     user = User.find(params[:user_id])
-    posts = user.posts.map do |post|
-      post.attributes.merge(image_url: post.image_url)
+    if user == current_user
+      posts = user.posts.map do |post|
+        post.attributes.merge(image_url: post.image_url)
+      end
+      render json: posts.to_json(include: :user), status: :ok
+    else
+      render json: { error: 'Not authorized' }, status: :unauthorized
     end
-    render json: posts.to_json(include: :user), status: :ok
   end
-  
+
   def show
     post = Post.find(params[:id])
-    render json: post.to_json(include: [:user]), status: :ok
+    if post.user == current_user
+      render json: post.to_json(include: [:user]), status: :ok
+    else
+      render json: { error: 'Not authorized' }, status: :unauthorized
+    end
   end
 
   def update
     post = Post.find(params[:id])
-    post.update!(post_params)
-    render json: post.to_json(include: [:user]), status: :accepted
+    if post.user == current_user
+      post.update!(post_params)
+      render json: post.to_json(include: [:user]), status: :accepted
+    else
+      render json: { error: 'Not authorized' }, status: :unauthorized
+    end
   end
   
   def create
@@ -32,15 +44,17 @@ class PostsController < ApplicationController
 
   def destroy
     post = Post.find(params[:id])
-    post.destroy
-    render json: post
+    if post.user == current_user
+      post.destroy
+      render json: post
+    else
+      render json: { error: 'Not authorized' }, status: :unauthorized
+    end
   end
 
-private
+  private
 
-def post_params
-  params.permit(:image, :description, :user_id)
-end
-
-
+  def post_params
+    params.permit(:image, :description, :user_id)
+  end
 end
